+++ App/App.js
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/lib/integration/react'
 import createStore from 'App/Stores'
 import RootScreen from './Containers/Root/RootScreen'
-import Chart from './Components/Chart.react.js'
+import Chart from './Components/Chart.js'
 
 const { store, persistor } = createStore()
