+++ App/Components/ChartStyle.js
@@ -1,10 +1,8 @@
 import { StyleSheet } from 'react-native'
+import { Helpers, Metrics, Fonts, Colors } from 'App/Theme'
 
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF'
-  }}
-)
+export default StyleSheet.create({
+  title:{
+    textAlign: 'center',
+  }
+})
